Allow for keeping succesful tests output directories if the V env variable is non-null.

+++ development/autotests/run-tests.sh
@@ -100,7 +100,7 @@ for tf in $(echo "$TESTS"); do
     if $cmd > test-log.txt 2>&1; then
 	echo Ok
 	cd ..
-	rm -rf "out-$t";
+	if [ "$V" != "" ]; then rm -rf "out-$t"; fi
     else
 	echo FAILED
 	cd ..
